Move the content API JSON type resolver to an appropriate namespace
