+++ src/Umbraco.Cms.Api.Content/Json/ContentApiJsonTypeResolver.cs
@@ -3,7 +3,7 @@ using System.Text.Json.Serialization;
 using System.Text.Json.Serialization.Metadata;
 using Umbraco.Cms.Core.Models.ContentApi;
 
-namespace Umbraco.Cms.Api.Content.Filters;
+namespace Umbraco.Cms.Api.Content.Json;
 
 // see https://learn.microsoft.com/en-us/dotnet/standard/serialization/system-text-json/polymorphism?pivots=dotnet-7-0
 // TODO: if this type resolver is to be used for extendable content models (custom IApiContent implementations) we need to work out an extension model for known derived types
